requirements.txt
kaggle
numpy
matplotlib
pandas
scikit-learn
python-box
pytorch-lightning
warmup_scheduler
opencv-python
segmentation_models_pytorch
albumentations
tqdm
python-dotenv

slack_sdk
ipykernel
black
autopep8